Divide the movie page from the main page.
