+++ lib/main.dart
@@ -2,9 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:movie_n_chat/pages/main_page.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -19,7 +21,7 @@ class MyApp extends StatelessWidget {
             .copyWith(background: Colors.grey[200])
             .copyWith(secondary: Colors.blueGrey),
       ),
-      home: const MyHomePage(),
+      home: const MainPage(),
     );
   }
 }
