+++ lib/pages/chat_page.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class ChatPage extends StatelessWidget {
+  const ChatPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Chat'),
+      ),
+      body: const Center(
+        child: Text('Chat Page'),
+      ),
+    );
+  }
+}
